LayoutTests/fast/profiler/anonymous-function-calls-built-in-functions.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
if (window.internals)
    internals.settings.setJavaScriptProfilingEnabled(true);

function startTest()
{
    console.profile("Anonymous function calls built-in functions");
    anonymousFunction();
    endTest();
}
</script>
</head>

<body onload="startTest()">
This page has an anonymous JavaScript function that calls built-in functions.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  Beneath onload and startTest it should show three children, an
anonymous function, script context, and the endTest call.
<div id="output"></div>
</body>
</html>
