LayoutTests/fast/profiler/anonymous-function-calls-built-in-functions-expected.txt

This page has an anonymous JavaScript function that calls built-in functions. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. Beneath onload and startTest it should show three children, an anonymous function, script context, and the endTest call.

Profile title: Anonymous function calls built-in functions
Thread_1 (no file) (line 0:0)
   startTest anonymous-function-calls-built-in-functions.html (line 11:1)
      anonymousFunction profiler-test-JS-resources.js (line 29:37)
         insertNewText profiler-test-JS-resources.js (line 17:26)
            createElement (no file) (line 0:0)
            createTextNode (no file) (line 0:0)
            appendChild (no file) (line 0:0)
            getElementById (no file) (line 0:0)
      endTest profiler-test-JS-resources.js (line 1:20)
